// src/api/authApi/core/_modales.ts
export type Data_Form = {
  store_name: string;
  store_email: string;
  store_phone: string;
  category: string;
  tax_registration_document: File | null;
  location: string;
  website_url: string;
  first_name: string;
  last_name: string;
  email: string;
  phone: string;
  password: string;
  confirmPassword: string;
  national_id_document: File | null;
};